Moved _test_private_date_setters elsewhere
