+++ sunpy/map/sources/tests/test_cor_source.py
@@ -7,7 +7,7 @@ import astropy.units as u
 
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.stereo import CORMap
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 __author__ = 'Pritish C. (VaticanCameos)'
