+++ sunpy/map/sources/tests/test_eit_source.py
@@ -8,7 +8,7 @@ import astropy.units as u
 import sunpy.map
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.soho import EITMap
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 __author__ = "Pritish C. (VaticanCameos)"
