+++ sunpy/map/sources/tests/test_gong_synoptic_source.py
@@ -4,7 +4,7 @@ import astropy.units as u
 
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.gong import GONGSynopticMap
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 
 @pytest.fixture
