+++ sunpy/map/sources/tests/test_kcor_source.py
@@ -8,7 +8,7 @@ from astropy.tests.helper import assert_quantity_allclose
 
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.mlso import KCorMap
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 
 @pytest.fixture()
