+++ sunpy/map/sources/tests/test_mdi_source.py
@@ -9,7 +9,7 @@ from astropy.coordinates import Angle
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.soho import MDIMap, MDISynopticMap
 from sunpy.util.exceptions import SunpyMetadataWarning
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 __author__ = 'Pritish C. (VaticanCameos)'
