+++ sunpy/map/sources/tests/test_rhessi_source.py
@@ -7,7 +7,7 @@ from sunpy.data.test import get_test_filepath
 from sunpy.map import Map
 from sunpy.map.sources.rhessi import RHESSIMap
 from sunpy.util.exceptions import SunpyMetadataWarning
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 
 @pytest.fixture(scope="module")
