+++ sunpy/map/sources/tests/test_xrt_source.py
@@ -9,7 +9,7 @@ import astropy.units as u
 
 from sunpy.data.test import get_dummy_map_from_header, get_test_filepath
 from sunpy.map.sources.hinode import XRTMap
-from . import _test_private_date_setters
+from .helpers import _test_private_date_setters
 
 __author__ = 'Pritish C. (VaticanCameos)'
